add criteria in index

+++ cf_xarray/tests/test_scripts.py
@@ -3,11 +3,24 @@ import os
 from cf_xarray.scripts import make_doc
 
 
+def remove_if_exist(paths):
+    for path in paths:
+        if os.path.exists(path):
+            os.remove(path)
+
+
 def test_make_doc():
 
-    path = "_build/csv/coordinate_criteria.csv"
+    # Create/remove files from tests/,
+    # always return to original working directory
+    owd = os.getcwd()
+    os.chdir(os.path.dirname(__file__))
     try:
+        names = ["axes", "coordinates", "coordinate_criteria"]
+        paths_to_check = [f"_build/csv/{name}.csv" for name in names]
+        remove_if_exist(paths_to_check)
+
         make_doc.main()
-        assert os.path.exists(path)
+        assert all(os.path.exists(path) for path in paths_to_check)
     finally:
-        os.remove(path)
+        os.chdir(owd)
